pyproject.toml
[tool.poetry]
name = "careersync"
version = "0.1.0"
description = ""
authors = ["Vyacheslav <slava20031112@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
uuid7 = "^0.1.0"
bcrypt = "^4.2.0"
passlib = "^1.7.4"
psycopg = "^3.2.2"
psycopg-binary = "^3.2.2"
psycopg2 = "^2.9.9"
invoke = "^2.2.0"
phonenumbers = "^8.13.46"
sqladmin = {extras = ["full"], version = "^0.19.0"}
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
